Add shared markdown renderer and update pages
